backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
alembic==1.13.1
python-dotenv==1.0.0
google-generativeai==0.3.2
pandas==2.1.4
numpy==1.26.3
python-multipart==0.0.6
aiosqlite==0.19.0
pydantic==2.5.3
